Finish login/register system
Formatted everything, and brought it up to programming standard.
Able to login and register account!
