+++ .idea/inspectionProfiles/Project_Default.xml
@@ -5,7 +5,7 @@
       <option name="minLines" value="5" />
     </inspection_tool>
     <inspection_tool class="ExtractMethodRecommender" enabled="true" level="WARNING" enabled_by_default="true">
-      <option name="minLength" value="725" />
+      <option name="minLength" value="1092" />
     </inspection_tool>
     <inspection_tool class="ThrowablePrintStackTrace" enabled="true" level="WARNING" enabled_by_default="true" editorAttributes="WARNING_ATTRIBUTES" />
   </profile>
