+++ src/constants/CommonConstants.java
@@ -2,7 +2,7 @@ package constants;
 
 import java.awt.*;
 
-public class CommonConstants {
+public final class CommonConstants {
     // color hex values
     public static final Color PRIMARY_COLOUR = Color.decode("#1C5630");
     public static final Color SECONDARY_COLOUR = Color.decode("#fcec03");
